+++ dna/models/torch_modules/hidden_mlp_dag_lstm_mlp.py
@@ -1,7 +1,6 @@
 import torch.nn as nn
 
-from . import ACTIVATIONS
-from . import PyTorchRandomStateContext
+from .torch_utils import PyTorchRandomStateContext, get_activation
 from .dag_lstm import DAGLSTM
 from .submodule import Submodule
 
@@ -32,7 +31,7 @@ class HiddenMLPDAGLSTMMLP(nn.Module):
                 input_mlp_layer_sizes, mlp_activation_name, mlp_use_batch_norm, mlp_use_skip, dropout,
                 device=self.device, seed=self._input_mlp_seed
             ),
-            ACTIVATIONS[mlp_activation_name]()
+            get_activation(mlp_activation_name)()
         ]
         if dropout > 0.0:
             input_dropout = nn.Dropout(p=dropout)
