+++ dna/models/torch_modules/lstm_mlp.py
@@ -1,7 +1,7 @@
 import torch
 import torch.nn as nn
 
-from . import PyTorchRandomStateContext
+from .torch_utils import PyTorchRandomStateContext
 from .submodule import Submodule
