+++ dna/models/torch_modules/pmf.py
@@ -1,7 +1,7 @@
 import torch
 import torch.nn as nn
 
-from . import PyTorchRandomStateContext
+from .torch_utils import PyTorchRandomStateContext
 
 
 class PMF(nn.Module):
